refactor: extract translate function from service to translate
