+++ src/ui/map/map.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import mapboxgl from 'mapbox-gl';
-import { translite } from "../../utils/service";
+import translite from '../../utils/translite';
 
 mapboxgl.accessToken = 'pk.eyJ1IjoieXVza292c3MiLCJhIjoiY2szNnRoODJuMDR2NTNubWhmZ2tyNHcweiJ9.nqpNmNuvXFtwx0VFoqqxGQ';
 
@@ -17,7 +17,7 @@ export class Map extends React.Component {
       container: this.mapContainer,
       center: [this.props.lon, this.props.lat],
       style: 'mapbox://styles/mapbox/streets-v9',
-      zoom: 10
+      zoom: 10,
     });
   }
 
@@ -35,18 +35,34 @@ export class Map extends React.Component {
       width: '100%',
       height: '100%',
     };
-    
-    const lat = (''+this.props.lat).split('.');
-    const lon = (''+this.props.lon).split('.');
+
+    const lat = (`${this.props.lat}`).split('.');
+    const lon = (`${this.props.lon}`).split('.');
 
     return (
       <div className="geoposition">
-        <div className='mapContainer'>
-          <div style={style} ref={el => this.mapContainer = el} className='map'/>
+        <div className="mapContainer">
+          <div style={style} ref={(el) => this.mapContainer = el} className="map" />
         </div>
-        <p>{translite(this.props.lang,'words', 'lat')}: { lat[0] }&deg;{ lat[1] }'</p>
-        <p>{translite(this.props.lang,'words', 'lon')}: { lon[0] }&deg;{ lon[1] }'</p>
+        <p>
+{translite(this.props.lang,'words', 'lat')}
+:
+{' '}
+{ lat[0] }
+&deg;
+{ lat[1] }
+'
+</p>
+        <p>
+{translite(this.props.lang,'words', 'lon')}
+:
+{' '}
+{ lon[0] }
+&deg;
+{ lon[1] }
+'
+</p>
       </div>
-    )
+    );
   }
 }
