+++ src/ui/search/cmp-search.js
@@ -1,25 +1,27 @@
 import React from 'react';
-import { translite } from '../../utils/service';
+import translite from '../../utils/translite';
 
-export default function Search(props){
+export default function Search(props) {
   return (
     <div className="search">
       <form className="search__form">
-        <input 
-          className="search__input" 
+        <input
+          className="search__input"
           type="text"
-          placeholder="Enter city name"
-          onChange={ props.onChange }
-          value={ props.value }
+          placeholder={translite(props.lang, 'words', 'searchPlaceholder')}
+          onChange={props.onChange}
+          value={props.value}
         />
-        <button 
+        <button
           className="button search__button"
           onClick={(event) => {
             props.onClick();
             event.preventDefault();
           }}
-        >{ translite(props.lang, 'words', 'search') }</button>
+        >
+{ translite(props.lang, 'words', 'search') }
+        </button>
       </form>
     </div>
-  )
+  );
 }
