+++ src/ui/weatherForecastDay/cmp-weatherForecastDay.js
@@ -1,16 +1,20 @@
 import React from 'react';
-import { translite, celsiusToFaringate } from '../../utils/service';
+import { celsiusToFaringate } from '../../utils/service';
+import translite from '../../utils/translite';
 
 export function WeatherForecastDay(props) {
   const { lang, forecast, isCelsius } = props;
   const temp = isCelsius ? forecast.temp : celsiusToFaringate(forecast.temp);
-  return(
-     <div className="weather-forecast__day">
-      <div className="weather-forecast__weekday">{ translite(lang,'fullDayName', forecast.weekDay) }</div>
-      <div className="weather-forecast__temp">{ temp }&deg;</div>
+  return (
+    <div className="weather-forecast__day">
+      <div className="weather-forecast__weekday">{ translite(lang, 'fullDayName', forecast.weekDay) }</div>
+      <div className="weather-forecast__temp">
+         { temp }
+&deg;
+              </div>
       <div className="weather-forecast__img">
-        <img src={`./images/${ forecast.icon }.png`} alt=''/>
-      </div>
+         <img src={`./images/${forecast.icon}.png`} alt="" />
+       </div>
     </div>
-  )
+  );
 }
